fix(jangar): split revenue repair parser helpers
