+++ docs/jangar/architecture-inventory.md
@@ -15,7 +15,7 @@ This file is generated by `bun run --cwd services/jangar docs:inventory:write`.
 | Rank | Module                                                                 |  LOC |
 | ---- | ---------------------------------------------------------------------- | ---: |
 | 1    | `services/jangar/src/server/supporting-primitives-controller.ts`       | 3303 |
-| 2    | `services/jangar/src/data/agents-control-plane.ts`                     | 2756 |
+| 2    | `services/jangar/src/data/agents-control-plane.ts`                     | 2775 |
 | 3    | `services/jangar/src/server/codex-judge.ts`                            | 2729 |
 | 4    | `services/jangar/src/server/atlas-store.ts`                            | 2148 |
 | 5    | `services/jangar/src/server/orchestration-controller.ts`               | 2140 |
